Update Coding_Challenge_5.py
Andy,

Thank you for your help last night, it definitely helped me get through it. 
One thing to note, for some reason the processing time gets exponentially longer when the cellSizeWidth/Height is lower than ~2. When it is 2, the processing takes about
~30 seconds on my machine, and then when I reduced it to 1, it was going for 10 minutes and had not yet stopped, but no errors were recorded, so it leads me to believe that it is just processing too much. The heatmap output with 2 cell size is still usable though, which is good. 

Also, thank you for the edit to help bring the header back into the CSV files. I had a really janky method that relied on inputting the entire header manually by using file.write(header), which is no good because hardcoding the header for one file does not work for any other files. I took that part out and kept your method, but just wanted to let you know it helped!